rename 2 submodules
