+++ pyproject.toml
@@ -47,7 +47,7 @@ preview = true
 [tool.ruff.lint]
 extend-select = ["I"]  # W505 (doc-line-too-long); D (pydocstyle)
 # TODO: Gradually unignore these.
-extend-ignore = ["N999", "DTZ005", "SIM115", "RUF012"]
+extend-ignore = ["DTZ005", "SIM115", "RUF012"]
 pydocstyle.convention = "numpy"
 pycodestyle.max-doc-length = 80
