+++ src/impala/superCal/__init__.py
@@ -1,4 +1,4 @@
 from . import post_process
 from .impala_clust import *
-from .impala_noProbit_emu import *
-from .models_withLik import *
+from .impala_noprobit_emu import *
+from .models_withlik import *
